pyproject.toml
[project]
name = "kinema"
version = "0.1.0"
description = "κίνημα — music video generator for a-u.supply"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.0",
    "PyYAML>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "requires_ffmpeg: requires ffmpeg installed",
    "requires_network: hits the a-u.supply API",
]
